+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to set application currency other than ADS
 - AdPanel's placeholders to config API
 - Allow change user rights
+- Login info to panel placeholders
 ### Changed
 - Split user access for advertisers and publishers
 - Config API returns data after store